+++ tests/data/supplementary/deleted-supplementary.csv
@@ -1,0 +1,1 @@
+bfed24dd-4184-461e-9cf7-01b3e0c82b05,37ae990f-6d65-499f-8fdc-ab89505cb52b
